requirements.txt
# Core dependencies
aiohttp>=3.9.0
aiofiles>=23.2.0
web3>=6.11.0
eth-utils>=2.3.0
websockets>=12.0

# CLI and output
rich>=13.7.0
tabulate>=0.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
orjson>=3.9.0
cachetools>=5.3.0
tenacity>=8.2.0
rate-limiter>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.5
aioresponses>=0.7.6

# Optional: FastAPI for web interface
fastapi>=0.109.0
uvicorn>=0.25.0

# Optional: GeckoTerminal API wrapper
# geckoterminal-api>=0.1.0
